+++ change-log.md
@@ -6,6 +6,7 @@
 - add: NotValueSource and not() expression
 - add: ExecuteJavascript step (Run script in browser)
 - add: DateFormatValueSource
+- add: NavigateForward, NavigateBack and RefreshPage steps
 - fix: ProjectResourceValueSource referencing a non-existent step fails with IllegalArgumentException (instead of the correct type)
 - fix: empty compound step fails (but should simply no-op)
 - upgrade to Selenium 2.53.0